fix: ensure cleanup happens properly
This PR updates the deletion portion of the control plane provider to
also trigger deletion of any control plane machines in the cluster. This
mimics the behavior done in the kubeadm control plane provider.
Additionally, there were some updates to the way we add ownerrefs to
make sure that the refs are created with blockOwnerDeletion set to true
to make sure the full cluster can't get torn down without tearing down
these decendents.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>