fix: ensure talosconfigs are created w/o controller=true in ownerref
This PR fixes a bug we were seeing where the ownerref that was getting
created by the taloscontrolplane was using a helper function and that
function set the controller: true in the ownerref. This keeps CAPI from
adding its own Machine ownerref to a given talosconfig because it
expects to be the controller for it.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>