fix: update initialization for CAPI v0.3.11
This PR updates the way we determine that our controlplanes are
initialized. We now say we're initialized if any nodes are registered
in the workload cluster.

This was needed because of an upstream change in moving to "phases" for
cluster api. We hit issues because we were saying that we were
initialized too early and that lead to a case where CAPI upstream would
never set the data secrets for a machine.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>