feat: audit and remove etcd members if machines no longer exist
This PR introduces some auditing functionality for etcd in the workload
clusters. If there are members in the workload etcd that don't
correspond to any CP node, the members will be removed. This helps in
the case where a machine is deleted out from under us and a new CP
machine comes up but can't join etcd.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>