fix: do not allow scaling down controlplane to zero
Doing that is pretty pointless as it's going to destroy a cluster.
Instead of doing that cluster should be deleted and recreated.

And with the new approach without init nodes scaling cluster back up
won't be able to succeed as bootstrap is done only a single time.

Still allowing user to set replicas to zero, but block scaling in the
controller with error messages when there's only a single node
remaining.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>